refactor: config 디렉토리 생성
기존의 SwaggerConfig, WebConfig 파일을 config 디렉토리 내로 이동했습니다.
