+++ movie-book/src/main/java/hello/moviebook/config/SwaggerConfig.java
@@ -1,4 +1,4 @@
-package hello.moviebook;
+package hello.moviebook.config;
 
 import io.swagger.v3.oas.models.Components;
 import io.swagger.v3.oas.models.OpenAPI;
