+++ movie-book/src/main/java/hello/moviebook/config/WebConfig.java
@@ -1,4 +1,4 @@
-package hello.moviebook;
+package hello.moviebook.config;
 
 import org.springframework.context.annotation.Configuration;
 import org.springframework.web.servlet.config.annotation.CorsRegistry;
